more 64bit fixes. Guile uses int and not size_t for the sizes of strings
Thu Oct 04 21:37:11 2001  George Lebl <jirka@5z.com>

	* ghtml.c: more 64bit fixes.  Guile uses int and not size_t for the
	  sizes of strings apparently.  It's kind of evil.

	* Makefile.am: fix make dist

	* journal.c: fix 64bit issue with saving

+++ gtt/ghtml.c
@@ -698,7 +698,7 @@ static SCM
 decode_scm_col_list (GttGhtml *ghtml, SCM col_list)
 {
 	SCM col_name;
-	size_t len;
+	int len;
 	char * tok = NULL;
 
 	/* reset the parser */
@@ -752,7 +752,7 @@ show_invoice (SCM col_list)
 static SCM
 do_show_scm (GttGhtml *ghtml, SCM node)
 {
-	size_t len;
+	int len;
 	char * str = NULL;
 
 	/* either a 'symbol or a "quoted string" */
